deps/v8/test/inspector/debugger/framework-stepping-expected.txt

Checks stepping with blackboxed frames on stack

Running test: testStepIntoFromUser
(anonymous) (expr.js:0:0)

Executing stepInto...
Executing stepInto...
userFoo (user.js:23:2)
frameworkCall (framework.js:10:23)
testStepFromUser (user.js:31:2)
(anonymous) (expr.js:0:0)

Executing stepInto...
Executing stepInto...
userBoo (user.js:27:2)
frameworkCall (framework.js:10:23)
testStepFromUser (user.js:31:2)
(anonymous) (expr.js:0:0)

Executing stepInto...
Executing stepInto...
testStepFromUser (user.js:32:0)
(anonymous) (expr.js:0:0)

Executing resume...

Running test: testStepOverFromUser
(anonymous) (expr.js:0:0)

Executing stepInto...
Executing stepInto...
userFoo (user.js:23:2)
frameworkCall (framework.js:10:23)
testStepFromUser (user.js:31:2)
(anonymous) (expr.js:0:0)

Executing stepOver...
Executing stepOver...
userBoo (user.js:27:2)
frameworkCall (framework.js:10:23)
testStepFromUser (user.js:31:2)
(anonymous) (expr.js:0:0)

Executing stepOver...
Executing stepOver...
testStepFromUser (user.js:32:0)
(anonymous) (expr.js:0:0)

Executing resume...

Running test: testStepOutFromUser
(anonymous) (expr.js:0:0)

Executing stepInto...
Executing stepInto...
userFoo (user.js:23:2)
frameworkCall (framework.js:10:23)
testStepFromUser (user.js:31:2)
(anonymous) (expr.js:0:0)

Executing stepOut...
userBoo (user.js:27:2)
frameworkCall (framework.js:10:23)
testStepFromUser (user.js:31:2)
(anonymous) (expr.js:0:0)

Executing resume...

Running test: testStepIntoFromFramework
frameworkBreakAndCall (framework.js:14:12)
testStepFromFramework (user.js:35:2)
(anonymous) (expr.js:0:0)

Executing stepInto...
userFoo (user.js:23:2)
frameworkBreakAndCall (framework.js:15:23)
testStepFromFramework (user.js:35:2)
(anonymous) (expr.js:0:0)

Executing resume...

Running test: testStepOverFromFramework
frameworkBreakAndCall (framework.js:14:12)
testStepFromFramework (user.js:35:2)
(anonymous) (expr.js:0:0)

Executing stepOver...
testStepFromFramework (user.js:36:0)
(anonymous) (expr.js:0:0)

Executing resume...

Running test: testStepOutFromFramework
frameworkBreakAndCall (framework.js:14:12)
testStepFromFramework (user.js:35:2)
(anonymous) (expr.js:0:0)

Executing stepOut...
testStepFromFramework (user.js:36:0)
(anonymous) (expr.js:0:0)

Executing resume...
